requirements.txt
beautifulsoup4==4.12.3
gradio==5.9.1
icalendar==5.0.13
python_dateutil==2.9.0
Requests==2.32.3
